docker_test/fish/views.py
from rest_framework import viewsets
from .models import Category, Fish
from .serializers import CategorySerializer, FishSerializer

class CategoryViewSet(viewsets.ModelViewSet):
    queryset = Category.objects.all()
    serializer_class = CategorySerializer

class FishViewSet(viewsets.ModelViewSet):
    queryset = Fish.objects.all()
    serializer_class = FishSerializer
