docker_test/requirements.txt
asgiref==3.8.1
Django==5.1.7
psycopg2-binary==2.9.10
python-environ==0.4.54
djangorestframework==3.15.2
sqlparse==0.5.3
pillow==11.1.0
django-cors-headers==4.7.0